Add makefile
